requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
pydantic==2.6.3
python-dotenv==1.0.1 